+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - style compilation documentation to readme
 - restrict bots from color searching
+- describe item's colors on primary image change
 
 ## [1.7.2] - 2019-02-27
 ### Added